feat(teknik-analiz): faz8a - coklu kirilim tarayicisi
- tlab/indicators/trend/breakouts.py::MultiBreakout -- TEK indikator, ~20
  kirilim turu (downtrend/uptrend/range/zone/hh/ll/n-week-high/MA/Donchian/
  Bollinger/kanal), confirm_bars parametreli, her biri icin retest_hold/
  false_break takip taramasi (ayni pattern_id ile zincirlenir, ORIJINAL
  kirilim kaydi asla degismez). Kalite skoru (hacim/seviye yasi/temas/
  govde/mesafe agirlikli).
- tlab/features/volatility.py::bollinger(), yeni tlab/features/channels.py
  ::regression_channel() -- Faz 2-EK'in yalnizca Faz 8A'nin ihtiyac duydugu
  iki parcasi (gerisi ayri bir takip isi).
- Iki gercek hata bulunup duzeltildi: price_structure.py'de trendline
  kirilim yonu (resistance/support) ters eslenmisti; IndicatorResult.
  to_json() numpy.bool_ tipini serialize edemiyordu (scanner'in surecler-
  arasi JSON aktarimiyla, gercek BIST verisiyle ortaya cikti).
- config/scans.yaml + tlab scan --preset dusen_kiran.
- 8 yeni test (211->219).

+++ teknik-analiz/tlab/indicators/structure/price_structure.py
@@ -173,7 +173,13 @@ def _trendlines(
                 )
             )
             if tl.broken_at is not None:
-                direction: Direction = "short" if kind == "resistance" else "long"
+                # build_trendlines'ın "beyond" tanımı: resistance (high
+                # pivotlarından çizilir) kırılımı close > line_val (YUKARI,
+                # boğa); support (low pivotlarından) kırılımı close < line_val
+                # (AŞAĞI, ayı). Önceki sürüm bunu ters atıyordu (Faz 8A'da
+                # breakouts.py yazılırken bulunan gerçek bir hata — hiçbir
+                # test direction alanını doğrulamıyordu).
+                direction: Direction = "long" if kind == "resistance" else "short"
                 signals.append(
                     Signal(
                         bar_time=df.index[tl.broken_at], detected_at=df.index[tl.broken_at],
